+++ tests/__init__.py
@@ -128,6 +128,7 @@ class DDSEndpoint:
     # Web
     INDEX = "/"
     LOGIN = "/login"
+    LOGOUT = "/logout"
     CANCEL_2FA = "/cancel_2fa"
     CONFIRM_2FA = "/confirm_2fa"
     CHANGE_PASSWORD = "/change_password"
@@ -190,7 +191,7 @@ class DDSEndpoint:
     PROJ_PUBLIC = BASE_ENDPOINT + "/proj/public"
     PROJ_PRIVATE = BASE_ENDPOINT + "/proj/private"
 
-    # Display facility usage
+    # Display unit usage
     USAGE = BASE_ENDPOINT + "/usage"
     INVOICE = BASE_ENDPOINT + "/invoice"